# api/swagger.yaml
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: acgw_session
  schemas:
...
    AdminDiscordRoleMapping:
      properties:
        discord_role_id:
          type: string
        role:
          type: string
      type: object
...
    AdminPermissionRequest:
      properties:
        permission:
          type: string
      type: object
    AdminRoleGrant:
      properties:
        permission:
          type: string
        role:
          type: string
      type: object
    AdminRolesResponse:
      properties:
        grants:
          items:
            $ref: '#/components/schemas/AdminRoleGrant'
          type: array
          uniqueItems: false
        mappings:
          items:
            $ref: '#/components/schemas/AdminDiscordRoleMapping'
          type: array
          uniqueItems: false
      type: object
...
    CallbackResponse:
      properties:
        discord_id:
          type: string
        roles:
          items:
            type: string
          type: array
          uniqueItems: false
        user_id:
          type: string
      type: object
...
    ErrorBody:
      properties:
        code:
          type: string
        details: {}
        message:
          type: string
      type: object
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        request_id:
          type: string
      type: object
...
    ListUsersResponse:
      properties:
        users:
          items:
            $ref: '#/components/schemas/User'
          type: array
          uniqueItems: false
      type: object
    MeResponse:
      properties:
        avatar:
          type: string
        created_at:
          type: string
        discord_id:
          type: string
        display_name:
          type: string
        global_name:
          type: string
        permissions:
          items:
            type: string
          type: array
          uniqueItems: false
        roles:
          items:
            type: string
          type: array
          uniqueItems: false
        user_id:
          type: string
        username:
          type: string
      type: object
...
    User:
      properties:
        created_at:
          type: string
        discord_id:
          type: string
        display_name:
          type: string
        global_name:
          type: string
        user_id:
          type: string
        username:
          type: string
      type: object
...
info:
  description: |-
    REST gateway between community applications and AzerothCore.
    It hides the AzerothCore SOAP/CLI internals behind a typed, authenticated and authorized HTTP API.
  title: ac-community-gw API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/admin/roles:
    get:
      description: Returns every role -> permission grant and Discord role mapping.
        Requires the gw.identity.roles.manage permission.
      operationId: identity.admin.roles.list
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRolesResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: List roles and mappings
      tags:
      - identity
  /api/v1/admin/roles/{role}/permissions:
    post:
      description: Grants a permission to a role (creating the role if needed). Requires
        the gw.identity.roles.manage permission.
      operationId: identity.admin.roles.grant
      parameters:
      - description: role
        in: path
        name: role
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPermissionRequest'
        description: permission
        required: true
      responses:
        "204":
          description: Granted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Grant a permission to a role
      tags:
      - identity
  /api/v1/admin/roles/{role}/permissions/{permission}:
    delete:
      description: Revokes a permission from a role. Requires the gw.identity.roles.manage
        permission.
      operationId: identity.admin.roles.revoke
      parameters:
      - description: role
        in: path
        name: role
        required: true
        schema:
          type: string
      - description: permission
        in: path
        name: permission
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Revoked
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Revoke a permission from a role
      tags:
      - identity
...
  /api/v1/auth/discord/callback:
    get:
      description: Consumes the single-use state, exchanges the code, provisions the
        user, synchronises roles and creates a session. Redirects to return_to or
        the configured post-login URL; returns the principal when neither is set.
      operationId: auth.discord.callback
      parameters:
      - description: authorization code
        in: query
        name: code
        required: true
        schema:
          type: string
      - description: single-use OAuth state
        in: query
        name: state
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackResponse'
          description: Returned when no redirect is configured
        "302":
          description: Redirect to return_to or the post-login URL
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Discord OAuth2 callback
      tags:
      - auth
  /api/v1/auth/discord/login:
    get:
      description: Mints a single-use state and PKCE verifier and redirects to Discord.
        The optional return_to must be a same-site absolute path.
      operationId: auth.discord.login
      parameters:
      - description: same-site absolute path to redirect to after login
        in: query
        name: return_to
        schema:
          type: string
      responses:
        "302":
          description: Redirect to Discord
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Start Discord login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Revokes the current server-side session and clears the session
        cookie.
      operationId: auth.logout
      responses:
        "204":
          description: Session revoked
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Log out
      tags:
      - auth
...
  /api/v1/identity/users:
    get:
      description: Lists provisioned community users. Requires the gw.identity.user.read
        permission.
      operationId: identity.users.list
      parameters:
      - description: filter by username or Discord name
        in: query
        name: filter
        schema:
          type: string
      - description: page size
        in: query
        name: limit
        schema:
          default: 50
          type: integer
      - description: page offset
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: List community users
      tags:
      - identity
  /api/v1/me:
    get:
      description: Returns the authenticated community user with their Discord profile,
        internal roles and effective permissions.
      operationId: auth.me
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Current principal
      tags:
      - auth
...